🎨 複雑化してきたのでメソッドを分けた
